Remove the static_link_warning from glibc (XCC)
If you build a program statically and use dlopen (or one of a few other
functions), you'll get a warning.  The warning is meant to tell you that
even though you built statically, you still need to have the right .so's
for things to work at runtime.  That's fine.

The problem is that, while this information is nice, it is a warning, and
not just informational, and it breaks with -Werror.

Rebuild your toolchain.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>