Throw errors from within tcpctl()
The #ip ctl message framework expects to get a string for an error, then
it just calls error() with that string.  We even go so far as to catch the
error, then return the current_errstr().  All a bit ridiculous.

Changing TCP's ctl message is the first in a few steps to change #ip's ctl
interface.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>