Remove the printk format %e
We were using it in two places, and one of them was incorrect (getchar()
wasn't returning an errno).

It was also blindly inferring errstr() from the context.  If we're going to
do that, we might as well not even pass in the err.  It was basically an
unused perror() for the kernel.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>