Remove user include hacks
Due to the old style of having user libraries include their own headers as
both <libname/foo.h> and <foo.h>, we had to have a few hacks to force us to
include the 'real' headers that we wanted.

Now that we do things the right way, we don't need to carry those hacks
around.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>