Demacro-ify the user BITMASK helpers
If you passed in e.g. bit 7 to CLR_BITMASK_ATOMIC(), the compiler would
give you a warning.  The joy of macros.

This implements most of the macros as inline functions.  Other than fixing
the bug with _size, the code should be the same as before - just as
functions, not macros.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>