Fix implicit declaration in procinfo.h (XCC)
The user-specific part of procinfo has a helper function that makes a
syscall.  That was implying __ros_syscall_errno() existed.  Ideally, we
wouldn't do that, but doing otherwise causes include loops.

Alternatively, we could just move or remove the functions.  Considering
they are for debugging, just externing in the function seems fine.

Reinstall your kernel headers.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>