Remove the O_NONBLOCK fcntl() intercept (XCC)
We need to change how nonblocking works, and make it a flag on the chan.
It's too much of a pain to have special casing for every device type -
let's just support fcntl and have O_NONBLOCK be one of its uses.

This will temporarily break non-blocking #ip conversations from glibc.

Rebuild glibc.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>