qio: Remove qconsume()
It's nasty and no one uses it (in Akaros).  It looks like the uart code
from Plan 9 used it.  This is the same deal as qproduce(), I think.

Specifically, qconsume() consumes up to len from the queue, but only from
the first non-empty block.

If we ever need this, we can bring it back or reimplement it with qget() or
something.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>