Use FD taps for event delivery for #alarm
It turns out that FD taps work for #alarm, and they do a better job than
the old ev_q.  Now the user can tell us what it wants to hear for a
particular tap (e.g. the alarmid) instead of the kernel having to imply
that.  Plus, now we use just the one mechanism for asking the kernel to
send to an event when something happens (FD tap).

This required a few changes to userspace.  The subtlest is that
devalarm_set_evq takes timerfd instead of ctlfd.  That's actually a sign
that the API might be too coupled to the implementation.  The nicer
cleanups in userspace are:

- tests/alarm.c and PVC alarms use the devalarm interfaces (those functions
  predated that interface)
- devalarm provides a way to extract the ID, instead of functions checking
  ev_msg->ev_arg2

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>