Change #alarm cancel method
Writing into a ctl is convenient, but its hard to get the status back out.
When you read ctl, you get the directory ID, not actual status.  We could
have a status variable that you parse, but that's a pain.

It's simpler to just have 0 > Qtimer mean "turn off the alarm".  That's
what this commit does, as well as removing the "cancel" command from Qctl.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>