Consolidate the two syscall tracers
The old syscall tracer is gone.  Now it all uses the same infrastructure.
You can still "trace syscall start" from the monitor, but that's it - it'll
trace everything.

This also fixes a couple minor things, which prompted the change.  First,
we were reporting a gibberish value for the retval on Entry records.  This
could be confusing in the traces.  Now they are just ---.  This required
separate format strings for entry and exit.

Also, on exit, we were setting the extra data to 0.  Say we came in on a
write; we'd capture the data then.  On exit, we'd lose it.  This might have
been intentional, and if so, I can change it back.  I found it a little
confusing when I was looking at Exit records.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>