Check notif_pending early in pop_user_ctx()
If notif_pending was set any time after being cleared (handle_events())
and before we fully popped the context, the pop would fail and we'd inch
our way down the uthread's stack.  If that happened every time (perhaps
due to a buggy 2LS, or a 2LS that receives regular events and does a
decent amount of processing during that window), then we'd eventually
run off the end of the stack in proc_pop_ctx.

The vulnerable window still exists, but it is shrunk down a lot.  You'd
have to get notif_pending set after the check *and* not before the
check.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>