Move ifaddrs to glibc (XCC)
Note that the file is ifaddrs.c in glibc, as compared to ifaddr.c.  The
sysdep has to match exactly what glibc expects.

We also need the libc_hidden_def for some reason.  Basically whatever
glibc does in other versions of ifaddrs.c needs to be done in our.

Rebuild glibc.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>