VMM: Add struct virtual_machine, use it in MMIO
As a whole, we'll need to track the overall state of the virtual machine.
That's what the struct virtual_machine is for.  Over time, we'll add more
bits to it, such as the memory mapping.

All VMM threads will know about their VM.  We'll need to track that
throughout the VMM.  MMIO creates threads, so it'll need to pass that
along.  (We might be able to infer it from current_uthread, but we're
already passing a guest_thread *).

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>