Add a parlib debug_print helper
Calling printf from vcore context is currently dangerous.  Both call into
glibc, which is what we want, but glibc's locking isn't PDR.  If a uthread
is doing a lot of printing, then we try to print from that vcore in a
handler, then we'll deadlock.  That needs to get fixed.  Until then, we
have this helper.  It also might be useful for other debugging.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>