Add a parlib control variable for yielding
This allows applications to control whether or not their 2LS yields.  You
might want to never yield for greedy performance or testing.

Note that the 2LS can ask for more vcores, if it wants/needs it.  It just
won't yield them.  This should be fine.  Apps can always ask for vcores
directly, which the 2LS won't yield.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>