Add a 2LS for VMMs
We now have a basic 2LS for VMMs.  It doesn't handle preemption or other
serious issues, but it does the basics.  Specifically, it supports
vmrunkernel, which is no longer a pthread-based app.

You can control whether vmrunkernel is an SCP or an MCP, as well as whether
or not it yields, with the parlib_ control variables.  We can provide
arguments to vmrunkernel for that later.

Ancillary changes:
- virtio uses the VMM 2LS, instead of pthreads
- renamed the apic() function (too generic)
- the vmexit switch code was moved out of vmrunkernel and refactored.  It
  should be easier to follow.  In the process, I removed the empty apic()
case and the unused interrupt-window case.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>