x86: Fix write_msr() casting
The hardware is supposed to ignore the upper 32 bits of rdx and rax, which
it does.  But we shouldn't pass in garbage.

Here's the asm before, with the input value in rax:

				# ecx previously loaded
	mov    %rax,%rdx 	# u64 rax has the val, putting the upper 32 in rdx
	shr    $0x20,%rdx 	# upper 32 in rdx
	wrmsr 			# rax still has the upper bits

Here's the new asm:

	mov    %rax,%rdx
	mov    %edi,%ecx
	mov    %eax,%eax 	# zeroes the upper part of rax
	shr    $0x20,%rdx 	# upper 32 in rdx
	wrmsr

When the compiler saw the u32 cast, it just ignores the masking.  However,
the inline asm actually takes rax, not eax.  So it's wrong to cast to a u32
there.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>