VMM: Fix MSR emulation (XCC)
emsr_ok() was busted.

First, it was swapping rax and rdx.  The functions take rdx first, but
rdmsr takes rax first.  This was the proximate cause of a bug where if you
ran a guest pcore that did Linux initialization on the same pcore that had
a previous VM do the same, then the kernel would GPF on wrmsr 0x1c9 <-
0x300000003.  Good times.

The rdmsr() macro also wasn't zeroing the upper part of rax.  So we'd have
old stuff floating around.  That would be okay if you only ever use it with
u32s, but since it's a macro, we didn't know that from its signature.
Great.

Finally, we were passing the upper 32 of rax through to write_msr or-ed
with rdx.  That's not supposed to happen either.

Reinstall your kernel header if you care.  Though that crappy macro
shouldn't be in a kernel header...

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>