VMM: Use safe MSR accessors
We clearly need write_msr_safe.  The guest could give us an unacceptable
value that could trigger a GPF.  For read, better safe than sorry.  It's
possible we add an MSR that we allow the guest to read, but that isn't on
every piece of hardware.  In that case, we could also trigger a GPF.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>