Attempt to merge VMRs during mprotect
If you do a bunch of mprotects on parts of an mmaped region with different
settings, you'd break it up.  But if we mprotected to make a collection of
regions the same, then we wouldn't merge the VMRs back.

You'd notice this with vmrunkernel - we'd have O(100) VMRs, many of the
adjacent VMRs had the same settings.  Now those ones are merged.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>