Report readablity/writablility via 9p stat
This extends 9p to have bits for readable and writable, in the O_NONBLOCK
sense.  This is required for select().

Pipes and #ip use qio helpers to set the bit.  Other devices will need to
do something similar, if they use queues.  If not, they'll have to do
something else.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>