Push control over vcores into parlib variables
We had an extension to pthreads: pthread_can_vcore_request().  It was a way
for an application to tell the 2LS to not request vcores or to yield
vcores.  The problem with making it a pthread extension is that it is 2LS
specific: an app needs to know/care about it's 2LS, and every 2LS needed to
reimplement the same logic.

By pushing it into parlib, we avoid all of that.  The app also gets
finer-grained control over what it needs (i.e., there's a minor difference
between not yielding and not requesting more cores).

Be careful using this new variable.  It'll prevent *any* vcores from being
requested, so you want to use it after the app requests whatever vcores it
wants (not including the one VC you get from being an MCP).  This variable
is a minor pain, and might not be worth keeping around.  We'll see.

This is actually a problem of having 2LSs that aren't app-specific enough.
The app knows what it wants, but the 2LS doesn't.  We're trying to find a
way to deal with that in a way that doesn't duplicate too much code.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>