Extract busybox's make helpers
Parts of busybox's Makefile are generic to all apps that we'll build.
Instead of just copying bits and pieces, we can have all of the apps
include a Makefrag.

It's far from perfect.  We could detect the cross compiler instead of being
told what it is, and if AKAROS_ROOT (e.g. from .bashrc) doesn't point to
the repo we're in, then weird things could happen.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>