perf: Set sample_period
Newer versions of perf need the sample_period.  Without it, all the
percentages in perf report are 0.

The sample period should be how many samples it takes to trigger an event
(e.g. overflow IRQ on a perf counter).  Perhaps we should always have been
saying at least sample_period = 1 (i.e. an event notif for every event).

We don't keep track of the actual period from perf's invocation.  We could
if we wanted to, but since percentages are relative, it's not a big deal at
this point.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>