Fix circular issues with event headers (XCC)
The mailbox headers (e.g. ucq.h) needed parts of event.h, but later parts
of event.h needed the mailbox headers.  Previously we dealt with this by
being careful with where we #included.  Commit df786ca456cb ("Replace
 #include guards with #pragma once.") broke that for evbitmap.h.

The right fix is to split out the parts of the header that the mailboxes
need into it's own header.  This should help avoid future issues with other
mailboxes.  FWIW, I don't know if we could keep the old style (even if I
wanted to) with the #pragma once, which is probably why that one was left
as an ifndef.

Reinstall your kernel headers.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>