Add `ipconfig` from Plan 9.
ipconfig is an IP stack configuration tool.  It acts
as both a frontend to the filesystem-based configuration
dance one does when configuring a Plan 9-derived IP
stack, as well as a DHCP client and IPv6 autoconfiguration
agent.

This change is a port of `ipconfig` to Akaros.  Changes
include reformatting the code to match Akaros kernel style,
replacing `alarm` with Akaros system calls, removing
Plan 9- and GNU-specific code and replacing with more
portable idioms, and generally cleaning up.  This has been
tested on hardware for static and DHCP configuration of IPv4
and static configuration of IPv6.  It will be called by
`/ifconfig` if present in the KFS.  Fixes #4.

Change-Id: I233969a8d8efca429a773c7a8667f0bb7e88e552
Signed-off-by: Dan Cross <crossd@gmail.com>
Issue: 4
Bug: 28668810
Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>