Use gnu-user command line options for gcc (XCC)
We're mostly like the gnu-user config.  We don't use their gcc/config
exactly, but we take most of the parts.  We weren't taking their command
line options though, which was a problem.

If you tried to build with -rdynamic, gcc would complain.
-Wl,--export-dynamic would work, and it was just the lack of option parsing
on our part.

Thanks to James Knight for the tip to look at the opts file.

Rebuild gcc.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>